Fix environment handling again
The following code would previously incorrectly work:
```
(defun f () (print x))
(let ((x 0)) (f))  ; prints 0
```
This is because in order to maintain global scope, the closure would
look in the current scope when executing. In reality, it should look in
the same scope in which it was created, but subsequent modifications to
that scope should be visible to the closure. Therefore, we need the
closure to be in an IORef, and local contexts need to allocate a new
context containing the same mappings as the enclosing scope.

Since we don't need to modify mappings by using State, we now
simply use Reader, making the local state modifications much simpler, as
they are now simply `local`.