Properly implement arbitrary symbols
Formerly, keywords had special parsing for arbitrary keywords in the
form `:|...|`. This is actually a generalization of arbitrary symbols of
the form `|...|`, and even further, of the form `...|...|...` and so
forth. This pushes the arbitrariness back into symbol parsing and
handling.

Note that this still does not allow identifiers beginning with numbers.